long long -> int64_t
commit 00b28894eb1d51d45c5dd0733314922f25232706
Merge: 48d7e86 35ae434
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Sun Feb 18 11:38:44 2024 -0600

    Merge branch 'dev' into int64

commit 48d7e86c5eaffb1a9acff1b4212587e9c6623744
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Sun Feb 18 11:37:19 2024 -0600

    fi

commit f312b285278e33d0dacad142192177843e1f4082
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Sun Feb 18 11:35:31 2024 -0600

    f

commit bd39a3091ec5cfdb21c199298336f53cc8224610
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Sun Feb 18 11:33:54 2024 -0600

    fix

commit 0de9f83185f566a5563b4c3c531d7897999bf16e
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Sun Feb 18 11:31:35 2024 -0600

    another fix

commit 133f447fd89664429dcf1f09569f407f3e037e9c
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Sun Feb 18 11:30:01 2024 -0600

    fix test

commit a6e8f884687967aa9e31a3e9a100ed0117fa193a
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Sun Feb 18 11:25:17 2024 -0600

    fix

commit a3a9c97cba467fd0b523d1f2f9254bd7a7696cb0
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Sun Feb 18 11:23:24 2024 -0600

    switch these too

commit 6e8ebaeadddbd31d8e35772638f58c78055f476b
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Sun Feb 18 11:19:41 2024 -0600

    switch

commit 3ee228e29c010b05b23e1aad962f948e5b6f3d61
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Sun Feb 18 11:18:29 2024 -0600

    not long long

+++ library/data_structures/rmq.hpp
@@ -2,7 +2,7 @@
 #pragma once
 /**
  * @code{.cpp}
-       vector<long long> a;
+       vector<int64_t> a;
        RMQ rmq(a, ranges::min); // -std=c++20
        RMQ rmq(a, [&](auto& x, auto& y) { return min(x, y); });
  * @endcode
