+++ library/data_structures/uncommon/disjoint_rmq.hpp
@@ -6,7 +6,7 @@
  * Disjoint RMQ is like normal RMQ except the 2 query ranges never overlap.
  * @code{.cpp}
        //usage for min and # of mins:
-       vector<pair<long long, int>> a; //initialize a[i].second = 1
+       vector<pair<int64_t, int>> a; //initialize a[i].second = 1
        disjoint_rmq rmq(a, [&](auto& x, auto& y) {
            if (x.first == y.first) return make_pair(x.first, x.second + y.second);
            return min(x, y);
