+++ library/data_structures/uncommon/dsu_restorable.hpp
@@ -8,7 +8,7 @@
 struct dsu_restorable {
 	int num_sets;
 	vector<int> p;
-	vector<long long> subtree;
+	vector<int64_t> subtree;
 	vector<optional<array<int, 3>>> st;
 	dsu_restorable(int n): num_sets(n), p(n, -1), subtree(n) {}
 	inline int find(int u) {
@@ -33,6 +33,6 @@ struct dsu_restorable {
 	}
 	inline int size(int u) {return -p[find(u)];}
 	inline bool same_set(int u, int v) {return find(u) == find(v);}
-	inline long long sum(int u) {return subtree[find(u)];}
+	inline int64_t sum(int u) {return subtree[find(u)];}
 	inline void add(int u, int d) {while (u >= 0) subtree[u] += d, u = p[u];}
 };
