+++ library/flow/min_cost_max_flow.hpp
@@ -4,7 +4,7 @@
  * @see https://e-maxx.ru/algo/min_cost_flow
  */
 struct mcmf {
-	using ll = long long;
+	using ll = int64_t;
 	struct edge {
 		int u, v;
 		ll cap, cost, flow;
