+++ library/math/fibonacci.hpp
@@ -17,7 +17,7 @@ const int mod = 998'244'353;
  * @time O(log n)
  * @space O(log n) for recursion stack
  */
-array<long long, 2> fib(long long n) {
+array<int64_t, 2> fib(int64_t n) {
 	if (n == 0) return {0LL, 1LL};
 	auto [x, y] = fib(n >> 1);
 	auto c = x * (2 * y - x + mod) % mod, d = (x * x + y * y) % mod;
