+++ library/math/matrix_related/row_reduce.hpp
@@ -11,7 +11,7 @@
  * @time O(n * m * min(cols, n))
  * @space besides the O(n * m) `mat` param, this function uses O(1) space
  */
-pair<int, long long> row_reduce(vector<vector<long long>>& mat, int cols) {
+pair<int, int64_t> row_reduce(vector<vector<int64_t>>& mat, int cols) {
 	int n = ssize(mat), m = ssize(mat[0]), rank = 0;
 	auto det = 1LL;
 	assert(cols <= m);
