+++ library/math/matrix_related/xor_basis_ordered.hpp
@@ -5,7 +5,7 @@ const int lg = 60;
  * @see https://codeforces.com/blog/entry/68953 https://github.com/ssk4988/Hackpack /blob/main/content/numerical/XORBasis.h
  * @code{.cpp}
        basis<int> b;
-       basis<long long> b;
+       basis<int64_t> b;
        basis<bitset<lg>> b;
  * @endcode
  * can get tiny speed boost by repeatedly jumping to next 1-bit in v with __lg
