+++ tests/library_checker_aizu_tests/data_structures/bit_rupq.test.cpp
@@ -7,8 +7,8 @@ int main() {
 	cin.tie(0)->sync_with_stdio(0);
 	int n, q;
 	cin >> n >> q;
-	bit_rupq<long long> bit_i_1(n);
-	bit_rupq<long long> bit_i_2(vector<long long>(n, 50));
+	bit_rupq<int64_t> bit_i_1(n);
+	bit_rupq<int64_t> bit_i_2(vector<int64_t>(n, 50));
 	while (q--) {
 		int type;
 		cin >> type;
@@ -27,7 +27,7 @@ int main() {
 			int idx;
 			cin >> idx;
 			idx--;
-			long long res = bit_i_1.get_index(idx);
+			int64_t res = bit_i_1.get_index(idx);
 			auto curr = bit_i_2.get_index(idx);
 			assert(res == curr - 50);
 			cout << res << '\n';
